Create plotter_test_driver.py
All this does is wait for the plotter to open up, and then send it ten random points of data.  It's not the most beautiful code but I wanted to prove that the plotter works reasonably well and that I understood sockets in general before we get to Bluetooth (or whatever the case may be).